+++ guillotina/tests/test_catalog.py
@@ -156,16 +156,16 @@ async def test_query_stored_json(container_requester):
 
         conn = requester.db.storage.read_conn
         result = await conn.fetch('''
-select json from objects
+select json from {0}
 where json->>'type_name' = 'Item'
 order by json->>'id'
-''')
+'''.format(requester.db.storage._objects_table_name))
         assert len(result) == 2
         assert json.loads(result[0]['json'])['id'] == 'item1'
         assert json.loads(result[1]['json'])['id'] == 'item2'
 
         result = await conn.fetch('''
-select json from objects
+select json from {0}
 where json->>'id' = 'item1'
-''')
+'''.format(requester.db.storage._objects_table_name))
         assert len(result) == 1